Handle logical comparison

+++ test/arithmetic.cr
@@ -4,5 +4,5 @@ println(x)
 y: int = 10 + 10 * 10 + 10
 println(y * 1000)
 
-z: int = (2 + 3) * 5
+z: float = (2 + 3) * 5.0
 println(z)
